feat: Add createUser API endpoint
The code changes include adding a new API endpoint `/create` in the `user.ts` file under the `netlify/functions/api/routes` directory. This endpoint allows users to create a new user by providing their email and full name. If the email and full name are not provided, an error will be returned.

Based on the recent user commits and repository commits, it seems that the purpose of these changes is to add the `createUser` functionality to the "moner minder" project, specifically for Netlify functions.

Please note that the commit message should not include any meta information like issue references, tags, or author names.