feat: Add recentTransactions API endpoint
The code changes include adding a new API endpoint `/recent-transactions` in the `specials.ts` file under the `netlify/functions/api/routes` directory. This endpoint allows users to retrieve the recent transactions based on their email and year. If the email or year is not provided, an error will be returned.